📝 Add docstrings to `feature/currencyExchange`

+++ Modules/ExchangeRates/database/migrations/2025_03_23_205003_create_exchange_rates_table.php
@@ -7,7 +7,17 @@ use Illuminate\Support\Facades\Schema;
 return new class extends Migration
 {
     /**
-     * Run the migrations.
+     * Create the "exchange_rates" table.
+     *
+     * This migration sets up the "exchange_rates" table with the following columns:
+     * - id: Auto-incrementing primary key.
+     * - type: A string indicating the type of exchange rate.
+     * - date: A date field for the exchange rate's reference date.
+     * - value: A float representing the exchange rate value.
+     * - currency: A string for the currency code.
+     * - base_currency: A string for the base currency code.
+     * - source: A string denoting the exchange rate source (defaults to 'NBP').
+     * - created_at & updated_at: Timestamps managed automatically by the framework.
      */
     public function up(): void
     {
@@ -15,7 +25,7 @@ return new class extends Migration
             $table->id();
             $table->string('type');
             $table->date('date');
-            $table->decimal('value', 10, 6);
+            $table->float('value');
             $table->string('currency');
             $table->string('base_currency');
             $table->string('source')->default('NBP');
@@ -24,7 +34,10 @@ return new class extends Migration
     }
 
     /**
-     * Reverse the migrations.
+     * Revert the migration by dropping the "exchange_rates" table if it exists.
+     *
+     * This method undoes the changes made in the up() method, ensuring that the database schema
+     * returns to its state prior to this migration.
      */
     public function down(): void
     {
